requirements.txt
pyserial ~= 3.4
pywifi ~= 1.1.12
psutil ~= 5.9.2
